ultima version

+++ src/main/java/com/uv/deeplab/Dto/DParametros.java
@@ -24,9 +24,9 @@ public class DParametros implements Serializable {
 
     private String paramCamara;
 
-    private Integer paramVelMax;
+    private Float paramVelMax;
 
-    private Integer paramVelAngMax;
+    private Float paramVelAngMax;
 
     private Boolean paramMode;
